Relocate setup-wizard directory for better colocation
The so101/ directory contained only a setup-wizard/ subdirectory and
nothing else. This flattens the structure from so101/setup-wizard/ into
setup-wizard/so101/, making the feature boundary clearer and preparing
for future robot-type subdirectories (e.g. setup-wizard/trossen/).

All internal relative imports resolve at the same depth so no path
changes were needed inside the moved files. Only the external consumer
(so101-setup.tsx) had its two import paths updated.
