+++ application/ui/src/features/robots/setup-wizard/so101/calibration-step.tsx
