+++ application/ui/src/features/robots/setup-wizard/so101/diagnostics-step.tsx
