+++ application/ui/src/features/robots/setup-wizard/so101/motor-setup-step.tsx
