+++ application/ui/src/features/robots/setup-wizard/so101/setup-robot-viewer.tsx
