+++ application/ui/src/features/robots/setup-wizard/so101/setup-wizard.module.scss
