+++ application/ui/src/features/robots/setup-wizard/so101/stepper.tsx
