+++ application/ui/src/features/robots/setup-wizard/so101/use-calibration-animations.ts
