+++ application/ui/src/features/robots/setup-wizard/so101/use-joint-highlight.ts
