+++ application/ui/src/features/robots/setup-wizard/so101/use-setup-websocket.ts
