+++ application/ui/src/features/robots/setup-wizard/so101/verification-step.tsx
